- Diabling documetation building on travis, docs are building on read-the-docs. Traivs logs are shorted and cleaner. - Makefile builds with new muparser. - Cmake builds with new muparser. - Ctest is passing.
