+++ .travis.yml
@@ -40,5 +40,5 @@ install:
 
 script:
     - make BUILD=debug USE_NEUROKIT=1 && make clean
-    - mkdir -p _build && cd _build && cmake -DDEBUG=ON -DWITH_DOC=ON .. && make
+    - mkdir -p _build && cd _build && cmake -DDEBUG=ON -DWITH_DOC=OFF .. && make
     - ctest --output-on-failure
