+++ external/muparser/CMakeLists.txt
@@ -1,3 +1,4 @@
+cmake_minimum_required(VERSION 2.6)
 file(GLOB files_SRC "src/*.cpp")
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
 add_library(muparser ${files_SRC})
