+++ external/muparser/include/muParser.h
@@ -45,7 +45,7 @@ namespace mu
 
     <small>
     (C) 2011 Ingo Berg<br>
-    muparser(at)gmx.de
+    muparser(at)beltoforion.de
     </small>
   */
   /* final */ class Parser : public ParserBase
@@ -99,6 +99,7 @@ namespace mu
     // Prefix operators
     // !!! Unary Minus is a MUST if you want to use negative signs !!!
     static value_type  UnaryMinus(value_type);
+    static value_type  UnaryPlus(value_type);
 
     // Functions with variable number of arguments
     static value_type Sum(const value_type*, int);  // sum
