+++ external/muparser/include/muParserBase.h
@@ -32,6 +32,7 @@
 #include <map>
 #include <memory>
 #include <locale>
+#include <limits.h>
 
 //--- Parser includes --------------------------------------------------------------------------
 #include "muParserDef.h"
@@ -85,7 +86,7 @@ private:
     typedef ParserToken<value_type, string_type> token_type;
 
     /** \brief Maximum number of threads spawned by OpenMP when using the bulk mode. */
-    static const int s_MaxNumOpenMPThreads = 4;
+    static const int s_MaxNumOpenMPThreads = 16;
 
  public:
 
@@ -220,7 +221,12 @@ private:
 
       virtual std::string do_grouping() const 
       { 
-        return std::string(1, m_nGroup); 
+		// fix for issue 4: https://code.google.com/p/muparser/issues/detail?id=4
+		// courtesy of Jens Bartsch
+		// original code:
+		//        return std::string(1, (char)m_nGroup); 
+		// new code:
+		return std::string(1, (char)(m_cThousandsSep > 0 ? m_nGroup : CHAR_MAX));
       }
 
     private:
