+++ external/muparser/include/muParserTemplateMagic.h
@@ -108,43 +108,6 @@ namespace mu
     static T Sign(T v)  { return (T)((v<0) ? -1 : (v>0) ? 1 : 0); }
     static T Pow(T v1, T v2) { return std::pow(v1, v2); }
   };
-
-  // Create (mostly) dummy math function definitions for integer types. They are mostly
-  // empty since they are not applicable for integer values.
-#define MAKE_MATH_DUMMY(TYPE)                    \
-  template<>                                     \
-  struct MathImpl<TYPE>                          \
-  {                                              \
-    static TYPE Sin(TYPE)          { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Cos(TYPE)          { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Tan(TYPE)          { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE ASin(TYPE)         { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE ACos(TYPE)         { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE ATan(TYPE)         { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE ATan2(TYPE, TYPE)  { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Sinh(TYPE)         { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Cosh(TYPE)         { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Tanh(TYPE)         { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE ASinh(TYPE)        { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE ACosh(TYPE)        { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE ATanh(TYPE)        { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Log(TYPE)          { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Log2(TYPE)         { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Log10(TYPE)        { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Exp(TYPE)          { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Abs(TYPE)          { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Sqrt(TYPE)         { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Rint(TYPE)         { throw ParserError(_T("unimplemented function.")); } \
-    static TYPE Sign(TYPE v)          { return (TYPE)((v<0) ? -1 : (v>0) ? 1 : 0);     } \
-    static TYPE Pow(TYPE v1, TYPE v2) { return (TYPE)std::pow((double)v1, (double)v2); } \
-  };
-
-  MAKE_MATH_DUMMY(char)
-  MAKE_MATH_DUMMY(short)
-  MAKE_MATH_DUMMY(int)
-  MAKE_MATH_DUMMY(long)
-
-#undef MAKE_MATH_DUMMY
 }
 
 #endif
