+++ external/muparser/lib/Readme.txt
@@ -1,0 +1,1 @@
+Here goes the libraries (both static and shared) for this component.
