+++ external/muparser/src/muParserBase.cpp
@@ -25,11 +25,10 @@
 
 #include "muParserBase.h"
 #include "muParserTemplateMagic.h"
-#include "../debug/current_function.hpp"
-#include "../debug/print_function.hpp"
 
 //--- Standard includes ------------------------------------------------------------------------
 #include <cassert>
+#include <algorithm>
 #include <cmath>
 #include <memory>
 #include <vector>
@@ -101,7 +100,7 @@ namespace mu
   //---------------------------------------------------------------------------
   /** \brief Copy constructor. 
 
-    Tha parser can be safely copy constructed but the bytecode is reset during
+    The parser can be safely copy constructed but the bytecode is reset during
     copy construction.
   */
   ParserBase::ParserBase(const ParserBase &a_Parser)
@@ -131,9 +130,9 @@ namespace mu
   {}
 
   //---------------------------------------------------------------------------
-  /** \brief Assignement operator. 
+  /** \brief Assignment operator. 
 
-    Implemented by calling Assign(a_Parser). Self assignement is suppressed.
+    Implemented by calling Assign(a_Parser). Self assignment is suppressed.
     \param a_Parser Object to copy to this.
     \return *this
     \throw nothrow
@@ -267,8 +266,6 @@ namespace mu
   */
   string_type ParserBase::GetVersion(EParserVersionInfo eInfo) const
   {
-    string_type sCompileTimeSettings;
-    
     stringstream_type ss;
 
     ss << MUP_VERSION;
@@ -316,7 +313,7 @@ namespace mu
   /** \brief Add a value parsing function. 
       
       When parsing an expression muParser tries to detect values in the expression
-      string using different valident callbacks. Thuis it's possible to parse
+      string using different valident callbacks. Thus it's possible to parse
       for hex values, binary values and floating point values. 
   */
   void ParserBase::AddValIdent(identfun_type a_pCallback)
@@ -367,7 +364,7 @@ namespace mu
   //---------------------------------------------------------------------------
   /** \brief Check if a name contains invalid characters. 
 
-      \throw ParserException if the name contains invalid charakters.
+      \throw ParserException if the name contains invalid characters.
   */
   void ParserBase::CheckOprt(const string_type &a_sName,
                              const ParserCallback &a_Callback,
@@ -389,7 +386,7 @@ namespace mu
   //---------------------------------------------------------------------------
   /** \brief Check if a name contains invalid characters. 
 
-      \throw ParserException if the name contains invalid charakters.
+      \throw ParserException if the name contains invalid characters.
   */
   void ParserBase::CheckName(const string_type &a_sName,
                              const string_type &a_szCharSet) const
@@ -581,8 +578,8 @@ namespace mu
 
     CheckName(a_strName, ValidNameChars());
     
-    m_vStringVarBuf.push_back(a_strVal);           // Store variable string in internal buffer
-    m_StrVarDef[a_strName] = m_vStringBuf.size();  // bind buffer index to variable name
+    m_vStringVarBuf.push_back(a_strVal);                // Store variable string in internal buffer
+    m_StrVarDef[a_strName] = m_vStringVarBuf.size()-1;  // bind buffer index to variable name
 
     ReInit();
   }
@@ -590,7 +587,7 @@ namespace mu
   //---------------------------------------------------------------------------
   /** \brief Add a user defined variable. 
       \param [in] a_sName the variable name
-      \param [in] a_pVar A pointer to the variable vaule.
+      \param [in] a_pVar A pointer to the variable value.
       \post Will reset the Parser to string parsing mode.
       \throw ParserException in case the name contains invalid signs or a_pVar is NULL.
   */
@@ -697,13 +694,13 @@ namespace mu
       m_pParseFormula = &ParserBase::ParseString;
       m_pTokenReader->IgnoreUndefVar(false);
     }
-    catch(exception_type &e)
+    catch(exception_type & /*e*/)
     {
       // Make sure to stay in string parse mode, dont call ReInit()
       // because it deletes the array with the used variables
       m_pParseFormula = &ParserBase::ParseString;
       m_pTokenReader->IgnoreUndefVar(false);
-      throw e;
+      throw;
     }
     
     return m_pTokenReader->GetUsedVar();
@@ -789,7 +786,7 @@ namespace mu
       \param iArgCount Number of Arguments actually gathered used only for multiarg functions.
       \post The result is pushed to the value stack
       \post The function token is removed from the stack
-      \throw exception_type if Argument count does not mach function requirements.
+      \throw exception_type if Argument count does not match function requirements.
   */
   void ParserBase::ApplyFunc( ParserStack<token_type> &a_stOpt,
                               ParserStack<token_type> &a_stVal, 
@@ -861,15 +858,6 @@ namespace mu
             Error(ecTOO_FEW_PARAMS, m_pTokenReader->GetPos(), funTok.GetAsString());
 
           m_vRPN.AddFun(funTok.GetFuncAddr(), (funTok.GetArgCount()==-1) ? -iArgNumerical : iArgNumerical);
-          break;
-    default:
-          stringstream ss;
-          ss << "In function: " << SIMPLE_CURRENT_FUNCTION << " case statment " 
-              << " has incomplete defination "
-              << " Case " << funTok.GetCode() << " is not handled. " << endl;
-#ifdef  DEBUG
-          dump(ss.str(), "WARN");
-#endif     /* -----  DEBUG  ----- */
           break;
     }
 
@@ -1054,7 +1042,12 @@ namespace mu
       case  cmLOR:  --sidx; Stack[sidx]  = Stack[sidx] || Stack[sidx+1]; continue;
 
       case  cmASSIGN: 
-            --sidx; Stack[sidx] = *pTok->Oprt.ptr = Stack[sidx+1]; continue;
+          // Bugfix for Bulkmode:
+          // for details see:
+          //    https://groups.google.com/forum/embed/?place=forum/muparser-dev&showsearch=true&showpopout=true&showtabs=false&parenturl=http://muparser.beltoforion.de/mup_forum.html&afterlogin&pli=1#!topic/muparser-dev/szgatgoHTws
+          --sidx; Stack[sidx] = *(pTok->Oprt.ptr + nOffset) = Stack[sidx + 1]; continue;
+          // original code:
+          //--sidx; Stack[sidx] = *pTok->Oprt.ptr = Stack[sidx+1]; continue;
 
       //case  cmBO:  // unused, listed for compiler optimization purposes
       //case  cmBC:
@@ -1168,16 +1161,6 @@ namespace mu
                 }
               }
 
-        //case  cmSTRING:
-        //case  cmOPRT_BIN:
-        //case  cmOPRT_POSTFIX:
-        //case  cmOPRT_INFIX:
-        //      MUP_FAIL(INVALID_CODE_IN_BYTECODE);
-        //      continue;
-
-        //case  cmEND:
-	       //     return Stack[m_nFinalResultIdx];  
-
         default:
               Error(ecINTERNAL_ERROR, 3);
               return 0;
@@ -1197,11 +1180,10 @@ namespace mu
     ParserStack<int> stArgCount;
     token_type opta, opt;  // for storing operators
     token_type val, tval;  // for storing value
-    string_type strBuf;    // buffer for string function arguments
 
     ReInit();
     
-    // The outermost counter counts the number of seperated items
+    // The outermost counter counts the number of separated items
     // such as in "a=10,b=20,c=c+a"
     stArgCount.push(1);
     
@@ -1226,7 +1208,7 @@ namespace mu
                 break;
 
         case cmVAL:
-		            stVal.push(opt);
+		        stVal.push(opt);
                 m_vRPN.AddVal( opt.GetVal() );
                 break;
 
@@ -1362,7 +1344,7 @@ namespace mu
                 break;
 
         //
-        // Last section contains functions and operators implicitely mapped to functions
+        // Last section contains functions and operators implicitly mapped to functions
         //
         case cmBO:
                 stArgCount.push(1);
@@ -1558,7 +1540,7 @@ namespace mu
   }
 
   //---------------------------------------------------------------------------
-  /** \brief Enable the dumping of bytecode amd stack content on the console. 
+  /** \brief Enable the dumping of bytecode and stack content on the console. 
       \param bDumpCmd Flag to enable dumping of the current bytecode to the console.
       \param bDumpStack Flag to enable dumping of the stack content is written to the console.
 
@@ -1679,12 +1661,12 @@ namespace mu
   }
 
   //------------------------------------------------------------------------------
-  /** \brief Evaluate an expression containing comma seperated subexpressions 
+  /** \brief Evaluate an expression containing comma separated subexpressions 
       \param [out] nStackSize The total number of results available
       \return Pointer to the array containing all expression results
 
-      This member function can be used to retriev all results of an expression
-      made up of multiple comma seperated subexpressions (i.e. "x+y,sin(x),cos(y)")
+      This member function can be used to retrieve all results of an expression
+      made up of multiple comma separated subexpressions (i.e. "x+y,sin(x),cos(y)")
   */
   value_type* ParserBase::Eval(int &nStackSize) const
   {
@@ -1698,8 +1680,8 @@ namespace mu
   //---------------------------------------------------------------------------
   /** \brief Return the number of results on the calculation stack. 
   
-    If the expression contains comma seperated subexpressions (i.e. "sin(y), x+y"). 
-    There mey be more than one return value. This function returns the number of 
+    If the expression contains comma separated subexpressions (i.e. "sin(y), x+y"). 
+    There may be more than one return value. This function returns the number of 
     available results.
   */
   int ParserBase::GetNumResults() const
@@ -1714,7 +1696,7 @@ namespace mu
     I consider it important that Calc is a const function.
     Due to caching operations Calc changes only the state of internal variables with one exception
     m_UsedVar this is reset during string parsing and accessible from the outside. Instead of making
-    Calc non const GetUsedVar is non const because it explicitely calls Eval() forcing this update. 
+    Calc non const GetUsedVar is non const because it explicitly calls Eval() forcing this update. 
 
     \pre A formula must be set.
     \pre Variables must have been set (if needed)
@@ -1731,6 +1713,16 @@ namespace mu
   //---------------------------------------------------------------------------
   void ParserBase::Eval(value_type *results, int nBulkSize)
   {
+/* <ibg 2014-09-24/> Commented because it is making a unit test impossible
+
+    // Parallelization does not make sense for fewer than 10000 computations 
+    // due to thread creation overhead. If the bulk size is below 2000
+    // computation is refused. 
+    if (nBulkSize<2000)
+    {
+      throw ParserError(ecUNREASONABLE_NUMBER_OF_COMPUTATIONS);
+    }
+*/
     CreateRPN();
 
     int i = 0;
@@ -1743,7 +1735,7 @@ namespace mu
     #endif
 
     int nMaxThreads = std::min(omp_get_max_threads(), s_MaxNumOpenMPThreads);
-    int nThreadID, ct=0;
+	int nThreadID = 0, ct = 0;
     omp_set_num_threads(nMaxThreads);
 
     #pragma omp parallel for schedule(static, nBulkSize/nMaxThreads) private(nThreadID)
