+++ external/muparser/src/muParserCallback.cpp
@@ -56,7 +56,7 @@ namespace mu
 
 
   //---------------------------------------------------------------------------
-  /** \brief Constructor for constructing funcstion callbacks taking two arguments. 
+  /** \brief Constructor for constructing function callbacks taking two arguments. 
       \throw nothrow
   */
   ParserCallback::ParserCallback(fun_type2 a_pFun, bool a_bAllowOpti)
@@ -72,7 +72,7 @@ namespace mu
   //---------------------------------------------------------------------------
   /** \brief Constructor for constructing binary operator callbacks. 
       \param a_pFun Pointer to a static function taking two arguments
-      \param a_bAllowOpti A flag indicating this funcation can be optimized
+      \param a_bAllowOpti A flag indicating this function can be optimized
       \param a_iPrec The operator precedence
       \param a_eOprtAsct The operators associativity
       \throw nothrow
@@ -204,7 +204,7 @@ namespace mu
 
 
   //---------------------------------------------------------------------------
-  /** \brief Constructor for constructing funcstion callbacks taking two arguments. 
+  /** \brief Constructor for constructing function callbacks taking two arguments. 
       \throw nothrow
   */
   ParserCallback::ParserCallback(bulkfun_type2 a_pFun, bool a_bAllowOpti)
