+++ external/muparser/src/muParserError.cpp
@@ -100,7 +100,8 @@ namespace mu
     m_vErrMsg[ecUNEXPECTED_CONDITIONAL] = _T("The \"$TOK$\" operator must be preceeded by a closing bracket.");
     m_vErrMsg[ecMISSING_ELSE_CLAUSE]    = _T("If-then-else operator is missing an else clause");
     m_vErrMsg[ecMISPLACED_COLON]        = _T("Misplaced colon at position $POS$");
-
+    m_vErrMsg[ecUNREASONABLE_NUMBER_OF_COMPUTATIONS] = _T("Number of computations to small for bulk mode. (Vectorisation overhead too costly)");
+    
     #if defined(_DEBUG)
       for (int i=0; i<ecCOUNT; ++i)
         if (!m_vErrMsg[i].length())
@@ -159,7 +160,7 @@ namespace mu
       \param [in] a_iErrc the error code.
       \param [in] sTok The token string related to this error.
       \param [in] sExpr The expression related to the error.
-      \param [in] a_iPos the position in the expression where the error occured. 
+      \param [in] a_iPos the position in the expression where the error occurred. 
   */
   ParserError::ParserError( EErrorCodes iErrc,
                             const string_type &sTok,
@@ -182,7 +183,7 @@ namespace mu
   //------------------------------------------------------------------------------
   /** \brief Construct an error object. 
       \param [in] iErrc the error code.
-      \param [in] iPos the position in the expression where the error occured. 
+      \param [in] iPos the position in the expression where the error occurred. 
       \param [in] sTok The token string related to this error.
   */
   ParserError::ParserError(EErrorCodes iErrc, int iPos, const string_type &sTok) 
@@ -252,7 +253,7 @@ namespace mu
   {}
 
   //------------------------------------------------------------------------------
-  /** \brief Replace all ocuurences of a substring with another string. 
+  /** \brief Replace all occurrences of a substring with another string. 
       \param strFind The string that shall be replaced.
       \param strReplaceWith The string that should be inserted instead of strFind
   */
@@ -315,7 +316,7 @@ namespace mu
 
     If the error is not related to a distinct position this will return -1
   */
-  std::size_t ParserError::GetPos() const
+  int ParserError::GetPos() const
   {
     return m_iPos;
   }
